Refactor: add service layer & improve domain separation
Introduced service layer for categories, products, and orders, moving business logic out of controllers and repositories. Controllers now depend on service interfaces for better abstraction and testability. Updated repository interfaces to use entity models and moved them to a dedicated namespace. Applied UnitOfWork pattern consistently with a new CompleteAsync method. Updated DTOs and models (e.g., decimal for price, IsAvailable property, InitialStock). Enhanced order creation to handle multiple items and stock checks. Removed redundant SaveChangesAsync from repositories. Improved modularity and separation of concerns across the codebase.

+++ E-Commerce-Web-API/E-Commerce-Web-API/Models/Product.cs
@@ -3,13 +3,13 @@
     public class Product
     {
         public int ID { get; set; }
-        public string Name { get; set; }=string.Empty;
-        public float Price { get; set; }
-        public string Description { get; set; }= string.Empty;
+        public string Name { get; set; } = string.Empty;
+        public decimal Price { get; set; }
+        public string Description { get; set; } = string.Empty;
         public int CategoryID { get; set; }
         public Category Category { get; set; } = null!;
 
-        public Stock Stock { get; set; }=null!;
+        public Stock Stock { get; set; } = null!;
 
         public ICollection<Review> Reviews { get; set; } = new List<Review>();
